Update to Omniverse Physics 109.0 and PhysX SDK 5.7.0

+++ physx/include/foundation/PxIO.h
@@ -22,7 +22,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-// Copyright (c) 2008-2025 NVIDIA Corporation. All rights reserved.
+// Copyright (c) 2008-2026 NVIDIA Corporation. All rights reserved.
 // Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
 // Copyright (c) 2001-2004 NovodeX AG. All rights reserved.
 
@@ -61,7 +61,7 @@ class PxInputStream
 	\return the number of bytes read from the stream.
 	*/
 
-	virtual uint32_t read(void* dest, uint32_t count) = 0;
+	virtual uint64_t read(void* dest, uint64_t count) = 0;
 
 	virtual ~PxInputStream()
 	{
@@ -83,7 +83,7 @@ class PxInputData : public PxInputStream
 	\return size in bytes of the input data
 	*/
 
-	virtual uint32_t getLength() const = 0;
+	virtual uint64_t getLength() const = 0;
 
 	/**
 	\brief seek to the given offset from the start of the data.
@@ -92,7 +92,7 @@ class PxInputData : public PxInputStream
 	seek(length);
 	*/
 
-	virtual void seek(uint32_t offset) = 0;
+	virtual void seek(uint64_t offset) = 0;
 
 	/**
 	\brief return the current offset from the start of the data
@@ -100,7 +100,7 @@ class PxInputData : public PxInputStream
 	\return the offset to seek to.
 	*/
 
-	virtual uint32_t tell() const = 0;
+	virtual uint64_t tell() const = 0;
 
 	virtual ~PxInputData()
 	{
@@ -125,7 +125,7 @@ class PxOutputStream
 	\return the number of bytes written to the stream by this call.
 	*/
 
-	virtual uint32_t write(const void* src, uint32_t count) = 0;
+	virtual uint64_t write(const void* src, uint64_t count) = 0;
 
 	virtual ~PxOutputStream()
 	{
