+++ physx/source/common/src/CmPtrTable.cpp
@@ -22,7 +22,7 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
-// Copyright (c) 2008-2025 NVIDIA Corporation. All rights reserved.
+// Copyright (c) 2008-2026 NVIDIA Corporation. All rights reserved.
 // Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
 // Copyright (c) 2001-2004 NovodeX AG. All rights reserved.  
 
@@ -137,8 +137,8 @@ void PtrTable::add(void* ptr, PtrTableStorageManager& sm)
 	{
 		PX_ASSERT(!mBufferUsed);
 
-		if(!mOwnsMemory)								// don't own the memory, must always alloc
-			realloc(0, PxNextPowerOfTwo(mCount), sm);	// we're guaranteed nextPowerOfTwo(x) > x
+		if(!mOwnsMemory)										// don't own the memory, must always alloc
+			realloc(0, PxNextPowerOfTwo((uint32_t)mCount), sm);	// we're guaranteed nextPowerOfTwo(x) > x
 
 		else if(PxIsPowerOfTwo(mCount))					// count is at implicit capacity, so realloc
 			realloc(mCount, PxU32(mCount)*2, sm);		// ... to next higher power of 2
